services/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
groq
python-dotenv
psycopg2-binary
